Adicionando o método de finalizar/entregar simulado (ainda meio feio, mas aparentemente funcionando bem).

+++ backend/src/main/java/br/ifsul/enemsim/entidades/relacionais/auxiliar/EstudanteHabilidadeId.java
@@ -20,6 +20,6 @@ public class EstudanteHabilidadeId implements Serializable {
 
 	private Integer estudanteId;
 	
-	private Integer habilidadeId;
+	private Byte habilidadeId;
 	
 }
